Working FeistelCipher code!

+++ ksu/cs7530/obscura/encryption/FeistelFFunction.java
@@ -2,5 +2,5 @@ package ksu.cs7530.obscura.encryption;
 
 public interface FeistelFFunction {
 
-    public int transform(int input, int key);
+    public int transform(int input, long key);
 }
